refactor todo item card as separate file
